FIX: Clicking on subject in admin email logs broken
This was a regression introduced in
f0069a4fab4dccddb1092002661fb0b8f898f28c where clicking on an email
log's subject link did not open a modal that displays the details of the
incoming email.

In this commit, the following changes were also made:

1. Make all `admin-email-logs/*` routes inherit from the same route.
   Having two different parent routes doing the same thing was probably
   something that was missed in the refactor previously.

1. Introduce system tests for the admin email logs as we have no
   end-to-end coverage before this change.
