+++ app/assets/javascripts/admin/addon/routes/admin-email-logs/bounced.js
@@ -1,24 +1,14 @@
 import { action } from "@ember/object";
-import { service } from "@ember/service";
 import { popupAjaxError } from "discourse/lib/ajax-error";
 import { i18n } from "discourse-i18n";
 import IncomingEmail from "admin/models/incoming-email";
-import AdminEmailLogs from "admin/routes/admin-email-logs";
-import IncomingEmailModal from "../../components/modal/incoming-email";
+import AdminEmailLogsRoute from "admin/routes/admin-email-logs";
 
-export default class AdminEmailLogsBouncedRoute extends AdminEmailLogs {
-  @service modal;
-
+export default class AdminEmailLogsBouncedRoute extends AdminEmailLogsRoute {
   status = "bounced";
 
   titleToken() {
     return i18n("admin.config.email_logs.sub_pages.bounced.title");
-  }
-
-  @action
-  async showIncomingEmail(id) {
-    const model = await this.loadFromBounced(id);
-    this.modal.show(IncomingEmailModal, { model });
   }
 
   @action
