+++ app/assets/javascripts/admin/addon/routes/admin-email-logs/sent.js
@@ -1,7 +1,7 @@
 import { i18n } from "discourse-i18n";
-import AdminEmailLogs from "admin/routes/admin-email-logs";
+import AdminEmailLogsRoute from "admin/routes/admin-email-logs";
 
-export default class AdminEmailLogsSentRoute extends AdminEmailLogs {
+export default class AdminEmailLogsSentRoute extends AdminEmailLogsRoute {
   status = "sent";
 
   titleToken() {
