+++ app/assets/javascripts/admin/addon/templates/admin-email-logs/received.gjs
@@ -2,6 +2,7 @@ import { fn } from "@ember/helper";
 import { on } from "@ember/modifier";
 import RouteTemplate from "ember-route-template";
 import formatDate from "discourse/helpers/format-date";
+import routeAction from "discourse/helpers/route-action";
 import EmailLogsList from "admin/components/email-logs-list";
 import IncomingEmail from "admin/models/incoming-email";
 
@@ -37,6 +38,7 @@ export default RouteTemplate(
       @sourceModel={{IncomingEmail}}
       @headers={{RECEIVED_HEADERS}}
       @filters={{RECEIVED_FILTERS}}
+      @onShowEmail={{routeAction "showIncomingEmail"}}
     >
       <:default
         as |emailLog ccThreshold sortWithAddressFilter handleShowIncomingEmail|
