+++ spec/system/page_objects/modals/admin_incoming_email_modal.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module PageObjects
+  module Modals
+    class AdminIncomingEmailModal < PageObjects::Modals::Base
+      MODAL_SELECTOR = ".admin-incoming-email-modal"
+    end
+  end
+end
